backend/src/requirements.txt
Flask
Flask-Cors
Werkzeug==2.2.3
PyJWT==2.6.0
psycopg2==2.9.3
python-dotenv==0.21.0
cloudinary==1.27.0
pandas==1.5.3
numpy==1.23.4
scikit-learn==1.0.2
